requirements.txt
PyMuPDF==1.23.14
google-generativeai==0.3.2
python-dotenv==1.0.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0